Moved blueprints outside the instrumentserver
